+++ crates/opsml_interfaces/src/model/base/mod.rs
@@ -1,9 +1,7 @@
 pub mod interface;
-pub mod sklearn;
 pub mod types;
 pub mod utils;
 
 pub use interface::*;
-pub use sklearn::*;
 pub use types::*;
 pub use utils::*;
